Initial commit: move materials into ODS Repository
This project is now being worked on as part of ODS Labs, at Open Data
Services. Given this, and the fact that it is still in its very early
stages, I have chosen to move it to a new repository rather than bring
over the old repo with all the history.

The history was short, messy, and the commit messages were not suitable
for a public, professional, project.